@@ -11,29 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.5",
-    "@angular/cdk": "~12.0.5",
-    "@angular/common": "~12.0.5",
-    "@angular/compiler": "~12.0.5",
-    "@angular/core": "~12.0.5",
-    "@angular/forms": "~12.0.5",
-    "@angular/material": "^12.0.5",
-    "@angular/material-moment-adapter": "^12.0.5",
-    "@angular/platform-browser": "~12.0.5",
-    "@angular/platform-browser-dynamic": "~12.0.5",
-    "@angular/router": "~12.0.5",
-    "@smartstocktz/core-libs": "^0.2.12",
-    "bfastjs": "~4.2.9",
-    "moment": "^2.29.1",
-    "ngx-bottom-nav": "^11.0.0",
+    "@angular/animations": "~12.1.1",
+    "@angular/cdk": "~12.1.1",
+    "@angular/common": "~12.1.1",
+    "@angular/compiler": "~12.1.1",
+    "@angular/core": "~12.1.1",
+    "@angular/forms": "~12.1.1",
+    "@angular/material": "^12.1.1",
+    "@angular/material-moment-adapter": "^12.1.1",
+    "@angular/platform-browser": "~12.1.1",
+    "@angular/platform-browser-dynamic": "~12.1.1",
+    "@angular/router": "~12.1.1",
+    "@smartstocktz/core-libs": "~0.2.15",
+    "bfastjs": "~4.5.1-alpha.3",
+    "comlink": "~4.3.1",
+    "crypto-hash": "~1.3.0",
+    "moment": "~2.29.1",
     "rxjs": "~6.6.0",
     "tslib": "^2.3.0",
+    "uuid": "~8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.5",
-    "@angular/cli": "~12.0.5",
-    "@angular/compiler-cli": "~12.0.5",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler-cli": "~12.1.1",
     "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
