@@ -1,119 +1,104 @@
-import { Component, Inject, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { CustomerState } from '../states/customer.state';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {MatSnackBar} from '@angular/material/snack-bar';
+import {CustomerState} from '../states/customer.state';
+import {CustomerModel} from '../models/customer.model';
 
 @Component({
   template: `
-       <div class="row" mat-dialog-content>
-           <mat-card-title>Create Customer</mat-card-title>
-
+    <div class="">
       <form [formGroup]="createCustomerForm" (ngSubmit)="createCustomer()" class="create-shop-form-container" style="margin-top: 10px">
         <mat-form-field appearance="" style="width:100%">
-          <mat-label>First Name</mat-label>
-          <input matInput formControlName="firstName" placeholder="First Name">
-          <mat-error>First Name is required</mat-error>
+          <mat-label>Display name</mat-label>
+          <input matInput formControlName="displayName" placeholder="Full name">
+          <mat-error>Your name is required</mat-error>
         </mat-form-field>
-
-        <mat-form-field appearance="" style="width:100%">
-          <mat-label>Second Name</mat-label>
-          <input matInput formControlName="secondName" placeholder="Second Name">
-          <mat-error>Second Name is required</mat-error>
-        </mat-form-field>
-
         <mat-form-field appearance="" style="width:100%">
           <mat-label>Phone Number</mat-label>
-          <input matInput formControlName="phone" placeholder="Phone Number">
+          <input type="tel" matInput formControlName="phone" placeholder="Phone Number">
           <mat-error>Phone Number required</mat-error>
         </mat-form-field>
-
         <mat-form-field appearance="" style="width:100%">
           <mat-label>Email</mat-label>
-          <input matInput formControlName="email" placeholder="Email">
+          <input type="email" matInput formControlName="email" placeholder="Email">
           <mat-error>Email is required</mat-error>
         </mat-form-field>
-
         <mat-form-field appearance="" style="width:100%">
           <mat-label>Company</mat-label>
           <input matInput formControlName="company" placeholder="Company">
           <mat-error>Company name required</mat-error>
         </mat-form-field>
-
         <mat-form-field appearance="" style="width:100%">
           <mat-label>TIN No</mat-label>
           <input matInput formControlName="tin" placeholder="TIN No">
           <mat-error>TIN No required</mat-error>
         </mat-form-field>
-
-        <div class="row">
-          <button style="width: 100%" [disabled]="createShopProgress" class="ft-button btn-block" color="primary" mat-raised-button>
-            Create Customer
-            <mat-progress-spinner style="display: inline-block" *ngIf="createShopProgress" mode="indeterminate"
-                                  color="primary" [diameter]="20"></mat-progress-spinner>
+        <div class="">
+          <button style="width: 100%" [disabled]="customerState.saveCustomerFlag | async"
+                  class="ft-button btn-block" color="primary" mat-raised-button>
+            {{updateMode ? 'Update Customer' : 'Create Customer'}}
+            <mat-progress-spinner style="display: inline-block"
+                                  *ngIf="customerState.saveCustomerFlag | async" mode="indeterminate"
+                                  color="primary" [diameter]="20">
+            </mat-progress-spinner>
           </button>
         </div>
       </form>
+      <div style="margin-top: 6px;">
+        <button class="btn-block" mat-button color="primary" (click)="closeDialog($event)">
+          Close
+        </button>
+      </div>
     </div>
-    <div style="margin-top: 6px">
-      <button class="btn-block" mat-button color="primary" (click)="closeDialog($event)">
-        Close
-      </button>
-    </div>
-    `,
-  selector: 'app-create-customer'
+  `,
+  selector: 'app-create-customer-form'
 })
 export class CreateCustomerComponent implements OnInit {
   createCustomerForm: FormGroup;
-  createShopProgress = false;
+  @Input() updateMode = false;
+  @Input() customer: CustomerModel = null;
+  @Output() done = new EventEmitter<CustomerModel>();
 
-  constructor(public dialogRef: MatDialogRef<CreateCustomerComponent>,
-              private readonly formBuilder: FormBuilder,
-              @Inject(MAT_DIALOG_DATA) public data: any,
-              private readonly snack: MatSnackBar,
-              private readonly customerState: CustomerState) {
+  constructor(public readonly formBuilder: FormBuilder,
+              public readonly snack: MatSnackBar,
+              public readonly customerState: CustomerState) {
 
   }
 
   ngOnInit() {
     this.createCustomerForm = this.formBuilder.group({
-      firstName: ['', [Validators.nullValidator, Validators.required]],
-      secondName: ['', [Validators.nullValidator, Validators.required]],
-      phone: ['', [Validators.nullValidator, Validators.required]],
-      email: ['', [Validators.nullValidator]],
-      company: ['', [Validators.nullValidator]],
-      tin: ['', [Validators.nullValidator]],
+      displayName: [this.customer?.displayName, [Validators.nullValidator, Validators.required]],
+      phone: [this.customer?.phone, [Validators.nullValidator, Validators.required]],
+      email: [this.customer?.email],
+      company: [this.customer?.company],
+      tin: [this.customer?.tin],
       returns: [[]],
     });
   }
 
   createCustomer() {
     if (this.createCustomerForm.valid) {
-      this.createShopProgress = true;
-
-      this.createCustomerForm.value.displayName = this.createCustomerForm.value.firstName;
-      this.customerState.saveCustomer(this.createCustomerForm.value).then((val) => {
-        this.createShopProgress = false;
+      this.createCustomerForm.value.id = this.customer?.id;
+      this.customerState.saveCustomer(this.createCustomerForm.value).then((_1) => {
         this.snack.open('Customer Created Successfully', 'Ok', {
-          duration: 3000
+          duration: 2000
         });
-        this.dialogRef.close(null);
-      }).catch( err => {
-        this.snack.open('Failed to create Customer', 'Cancel', {
-          duration: 3000
+        this.done.emit(_1);
+      }).catch(_4 => {
+        this.snack.open(`Failed to ${this.updateMode ? 'update' : 'create'} customer`, 'Ok', {
+          duration: 2000
         });
       });
-
     } else {
       this.snack.open('Please fill all required fields', 'Ok', {
-        duration: 3000
+        duration: 2000
       });
     }
   }
 
   closeDialog($event: Event): void {
     $event.preventDefault();
-    this.dialogRef.close(null);
+    this.done.emit(null);
   }
 
 }
