@@ -10,7 +10,7 @@ import {InvoiceModel} from '../models/invoice.model';
       <div class="row header text-white align-items-center p-3">
         <div class="col-6 header-icon">
           <mat-icon class="ml-auto p-3">shopping_invoice</mat-icon>
-          <p >Invoice Details</p>
+          <p>Invoice Details</p>
         </div>
         <div class="col-6 text-right">
           <h3 class="mb-0">{{data.businessName}}</h3>
@@ -19,14 +19,14 @@ import {InvoiceModel} from '../models/invoice.model';
         </div>
       </div>
       <div class="row px-3 pt-4 m-0 justify-content-between">
-       <div>
-         <p class="mb-0">Invoice No.</p>
-         <p>{{data.id}}</p>
-       </div>
-       <div>
-         <p class="mb-0">Seller</p>
-         <p>{{data.sellerFirstName | titlecase }} {{data.sellerLastName | titlecase}}</p>
-       </div>
+        <div>
+          <p class="mb-0">Invoice No.</p>
+          <p>{{data.id}}</p>
+        </div>
+        <div>
+          <p class="mb-0">Seller</p>
+          <p>{{data.sellerFirstName | titlecase }} {{data.sellerLastName | titlecase}}</p>
+        </div>
       </div>
       <div class="row px-3 pt-4 m-0 justify-content-between">
         <div>
@@ -60,13 +60,14 @@ import {InvoiceModel} from '../models/invoice.model';
           </ng-container>
 
           <tr mat-header-row *matHeaderRowDef="invoiceDataColumns"></tr>
-          <tr class="table-data-row"  mat-row
+          <tr class="table-data-row" mat-row
               *matRowDef="let row; columns: invoiceDataColumns;"></tr>
 
         </table>
         <div class="d-flex pt-4 align-items-center justify-content-between">
           <h3 class="text-center col-4 ">Total Amount</h3>
-          <h2 class="text-white py-3 col-7 col-md-5 col-lg-6 text-center" style="background: #1b5e20;">{{data.amount | currency: ' '}} /=</h2>
+          <h2 class="text-white py-3 col-7 col-md-5 col-lg-6 text-center" style="background: #1b5e20;">{{data.amount | currency: ' '}}
+            /=</h2>
         </div>
       </div>
       <div class="py-3">
@@ -83,13 +84,14 @@ import {InvoiceModel} from '../models/invoice.model';
           </ng-container>
 
           <tr mat-header-row *matHeaderRowDef="returnsDataColumns"></tr>
-          <tr class="table-data-row"  mat-row
+          <tr class="table-data-row" mat-row
               *matRowDef="let row; columns: returnsDataColumns;"></tr>
 
         </table>
         <div class="d-flex pt-4 align-items-center justify-content-between">
           <h3 class="text-center col-4 ">Total Returns</h3>
-          <h2 class="text-white py-3 col-7 col-md-5 col-lg-6 text-center" style="background: #1b5e20;">{{totalAmount | currency: ' '}} /=</h2>
+          <h2 class="text-white py-3 col-7 col-md-5 col-lg-6 text-center" style="background: #1b5e20;">{{totalAmount | currency: ' '}}
+            /=</h2>
         </div>
       </div>
       <p class="text-center" style="color: #1b5e20">smartstock.co.tz</p>
@@ -97,7 +99,7 @@ import {InvoiceModel} from '../models/invoice.model';
   `,
   styleUrls: ['../styles/invoice-details.style.css'],
 })
-export class InvoiceDetailsComponent implements OnInit{
+export class InvoiceDetailsComponent implements OnInit {
   returnsData: MatTableDataSource<any>;
   returnsDataColumns = ['date', 'amount'];
   totalAmount = 0;
