@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree} from '@angular/router';
 import {Observable} from 'rxjs';
-import { UserService } from '../user-modules/user.service';
+import {UserService} from '@smartstocktz/core-libs';
 
 @Injectable({
   providedIn: 'root'
