@@ -2,27 +2,41 @@ import {StockModel} from './stock.model';
 import {CustomerModel} from './customer.model';
 
 export interface SalesModel {
-  isReturned?: boolean;
-  soldBy?: { username: string };
+  soldBy: {
+    username: string;
+  };
   id?: string;
+  _id?: string;
   idTra?: string;
-  date?: any;
-  product?: string;
-  category?: string;
-  unit?: string;
-  quantity?: number;
-  amount?: number;
-  customer?: string;
-  customerObject?: CustomerModel;
-  cartId?: string; // for retrieve sold items per cart/order
-  discount?: number;
-  user?: string;
-  sellerObject?: any;
-  timer: string;
-  channel?: string;
-  stock?: StockModel;
-  batch?: string; // for offline sync
+  createdAt?: any;
+  updatedAt?: any;
+  updatedBy?: any;
+  date: any;
+  timer: any;
+  product: string;
+  category: string;
+  unit: string;
+  quantity: number;
+  amount: number;
+  customer: string;
+  cartId?: string;
+  discount: number;
+  user: string;
+  sellerObject: {
+    username: string;
+    lastname: string;
+    firstname: string;
+    email: string;
+  };
+  channel: 'whole' | 'retail' | 'invoice' | 'online' | string;
+  stock: StockModel;
+  batch?: string;
   stockId: string;
-  creditor?: any;
-  paid?: boolean;
+  customerObject: {
+    firstName: string;
+    displayName: string;
+    lastName: string;
+    mobile: string;
+    email: string;
+  };
 }
