@@ -1,28 +1,33 @@
 export interface StockModel {
   createdAt?: any;
   updatedAt?: any;
-  barcode?: string;
   image?: any;
+  images?: any[];
+  barcode?: string;
   id?: string;
   _id?: string;
   product: string;
-  saleable: boolean | true;
-  canExpire?: boolean | true;
   description?: string;
+  type: 'simple' | 'grouped';
+  downloadable?: boolean;
+  downloads?: {
+    name: string;
+    type: string;
+    url: any;
+  }[];
+  retailPrice: number;
+  wholesalePrice: any;
+  catalog?: any[];
+  saleable: boolean;
+  canExpire: boolean;
   unit: string;
   category: string;
-  type?: 'simple' | 'grouped';
-  downloadable: boolean | false;
-  downloads: { name: string, type: string, url: string }[];
   stockable: boolean | true;
   purchasable: boolean | true;
   quantity: number;
   wholesaleQuantity: number;
   reorder: number;
   purchase: number;
-  retailPrice: number;
-  wholesalePrice: any;
   supplier: string;
   expire: string;
-  
 }
