@@ -1,7 +1,10 @@
 import {Component, OnInit} from '@angular/core';
-import {CreateCustomerComponent} from '../components/create-customer-form.component';
 import {MatDialog} from '@angular/material/dialog';
 import {DeviceState} from '@smartstocktz/core-libs';
+import {CustomerState} from '../states/customer.state';
+import {DialogCreateCustomerComponent} from '../components/dialog-create-customer.component';
+import {SheetCreateCustomerComponent} from '../components/sheet-create-customer.component';
+import {MatBottomSheet} from '@angular/material/bottom-sheet';
 
 @Component({
   selector: 'app-customer-page',
@@ -13,44 +16,71 @@ import {DeviceState} from '@smartstocktz/core-libs';
                         [body]="body"
                         backLink="/sale"
                         [hasBackRoute]="true"
+                        [showSearch]="true"
+                        (searchCallback)="onCustomerSearch($event)"
                         [leftDrawer]="side">
       <ng-template #side>
         <app-drawer></app-drawer>
       </ng-template>
       <ng-template #body>
-        <div class="container col-xl-9 col-lg-9 col-sm-10 col-md-10 col-sm-12 col-12 pt-3" style="min-height: 100vh">
-          <div class="d-flex flex-row align-items-center justify-content-between">
-            <button (click)="addCustomer()" color="primary" mat-flat-button
-                    class="m-2">
-              Add Customer
+        <div [class]="(deviceState.isSmallScreen | async)===true?'customers-container-mobile':'customers-container'">
+          <div class="actions-container">
+            <button (click)="addCustomer()" color="primary" mat-button>
+              <mat-icon matPrefix>add</mat-icon>
+              Add
             </button>
+            <button (click)="hotReload()" color="primary" mat-button>
+              <mat-icon matPrefix>refresh</mat-icon>
+              Load
+            </button>
+            <span class="actions-spacer"></span>
+            <div *ngIf="(deviceState.isSmallScreen | async)===false">
+              <mat-paginator #c_paginator></mat-paginator>
+            </div>
           </div>
-          <app-customer-list></app-customer-list>
+          <app-customer-list
+            [paginator]="(deviceState.isSmallScreen | async)===false?null:c_paginator"></app-customer-list>
         </div>
       </ng-template>
     </app-layout-sidenav>
   `,
-  styleUrls: []
+  styleUrls: ['../styles/customers-page.style.scss']
 })
 export class CustomersPage implements OnInit {
   constructor(private readonly dialog: MatDialog,
+              public readonly customerState: CustomerState,
+              public readonly matBottomSheet: MatBottomSheet,
               public readonly deviceState: DeviceState) {
   }
 
   async ngOnInit(): Promise<void> {
   }
 
-  async onSearch($event: string): Promise<void> {
-
-  }
-
-  async addCustomer(): Promise<any> {
-    const dialogRef = this.dialog.open(CreateCustomerComponent, {
-      // height: '400px',
-      // width: '600px',
+  addCustomer(): void {
+    if (this.deviceState.isSmallScreen.value === true) {
+      this.matBottomSheet.open(SheetCreateCustomerComponent, {
+        data: {},
+        closeOnNavigation: true
+      });
+      return;
+    }
+    const dialogRef = this.dialog.open(DialogCreateCustomerComponent, {
+      maxWidth: '500px'
     });
     dialogRef.afterClosed().subscribe(result => {
       console.log(`Dialog result: ${result}`);
     });
+  }
+
+  onCustomerSearch(query: string) {
+    if (!query || query === '') {
+      this.customerState.fetchCustomers();
+    } else {
+      this.customerState.search(query);
+    }
+  }
+
+  hotReload() {
+    this.customerState.hotFetchCustomers();
   }
 }
