@@ -1,24 +1,54 @@
-import { Injectable } from '@angular/core';
-import { StorageService } from '@smartstocktz/core-libs';
-import { BFast } from 'bfastjs';
+import {Injectable} from '@angular/core';
+import {UserService} from '@smartstocktz/core-libs';
+import {CustomerModel} from '../models/customer.model';
+import {wrap} from 'comlink';
+import {ShopModel} from '@smartstocktz/core-libs/models/shop.model';
+import {CustomerWorker} from '../workers/customer.worker';
 
 @Injectable({
-    'providedIn': 'root'
+  providedIn: 'root'
 })
-export class CustomerService{
-    constructor(private readonly storageService: StorageService) {
+export class CustomerService {
+  private customerWorker: CustomerWorker;
+
+  constructor(private readonly userService: UserService) {
+
+  }
+
+  private async initClass(shop: ShopModel) {
+    if (!this.customerWorker) {
+      const CW = wrap(new Worker(new URL('../workers/customer.worker', import.meta.url))) as unknown as any;
+      this.customerWorker = await new CW(shop);
     }
+  }
 
-    async getCustomers(size = 20, skip = 0): Promise<any[]> {
-        const shop = await this.storageService.getActiveShop();
-        const customers = await BFast.database(shop.projectId).collection('returns')
-          .query()
-          .skip(skip)
-          .size(size)
-          .orderBy('_created_at', -1)
-          .find<any[]>();
-        return customers;
-      }
+  async getCustomers(): Promise<CustomerModel[]> {
+    const shop = await this.userService.getCurrentShop();
+    await this.initClass(shop);
+    return this.customerWorker.getCustomers(shop);
+  }
 
+  async getRemoteCustomers(): Promise<CustomerModel[]> {
+    const shop = await this.userService.getCurrentShop();
+    await this.initClass(shop);
+    return this.customerWorker.getCustomersRemote(shop);
+  }
 
+  async createCustomer(customer: CustomerModel): Promise<CustomerModel> {
+    const shop = await this.userService.getCurrentShop();
+    await this.initClass(shop);
+    return this.customerWorker.createCustomer(customer, shop);
+  }
+
+  async search(query: string): Promise<CustomerModel[]> {
+    const shop = await this.userService.getCurrentShop();
+    await this.initClass(shop);
+    return this.customerWorker.search(query, shop);
+  }
+
+  async deleteCustomer(customer: CustomerModel): Promise<any> {
+    const shop = await this.userService.getCurrentShop();
+    await this.initClass(shop);
+    return this.customerWorker.deleteCustomer(customer, shop);
+  }
 }
