@@ -1,7 +1,5 @@
 import {Injectable} from '@angular/core';
-import { StorageService } from '@smartstocktz/core-libs';
-import {CustomerModel} from '../models/customer.model';
-import { CreditorService } from '../services/creditor.services';
+import {CreditorService} from '../services/creditor.services';
 
 @Injectable({
   providedIn: 'root'
